Setting Up the Homepage and Contact Page
Sorting out some base functionality and getting some CSS to start customizing the visuals of the page. Can also attach links to images so one of the main things to do soon is to create links to my GitHub and LinkedIn and use their logos as buttons to make it look nicer.