Adding more Content and tweaking with some CSS
The CSS definitely needs to be approved upon. For some reason I cant get the changes to show up when put into the browser, so I need to figure out exactly how to do that. The main thing to do is create more functionality to the website and come back to adding in pictures and organising it to later potentially. Not sure if this will prolong the time I have to spend on this project or not though.