Struggling a lot with wrapping my head around JavaScript but making progress in implementing more requirements for the portfolio
Struggled and procrastinated javascript today and just to make sure i got some progress in have added more content to the home page such as work experience and education sections just to tick them off in the project brief. Also managed to implement some more of the tags that they wanted us to use in the project as well and made a relevant table on google sheets to keep track of what aspects still need to be implemented by the time the project is finished. This will be my last update of the week due to a lack of time between work shifts on the weekend.